+++ CHANGELOG.md
@@ -12,6 +12,12 @@
   hostname, so that if it's possible each `CoreDNS` pods will sit on different infra node
   (PR[#3579](https://github.com/scality/metalk8s/pull/3579))
 
+- Allow to manage number of replicas and, soft and hard `podAntiAffinity`
+  for Control Plane Ingress Controller from Bootstrap configuration file, with
+  a default soft anti-affinity on hostname, so that if it's possible each
+  Control Plane Ingress Controller pods will sit on a different master node
+  (PR[#3617](https://github.com/scality/metalk8s/pull/3617))
+
 - Allow to manage soft and hard `podAntiAffinity` for `Dex` from Cluster
   and Services Configurations, with a default soft anti-affinity on hostname,
   so that if it's possible each `Dex` pods will sit on a different infra node